requirements-dev.txt
mypy
django-stubs[compatible-mypy]
pandas-stubs
python-dotenv